third party drop down api completed

+++ src/main/java/com/efitops/basesetup/dto/GrnDTO.java
@@ -1,10 +1,8 @@
 package com.efitops.basesetup.dto;
 
+import java.math.BigDecimal;
 import java.sql.Date;
-import java.time.LocalDate;
 import java.util.List;
-
-import javax.persistence.Column;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -24,7 +22,7 @@ public class GrnDTO {
 	private String gstType; 
 	private String adress;
 	private String currency;
-	private Long exchangeRate;
+	private BigDecimal exchangeRate;
 	private String grnClearTime;
 	private String invDcNo;
 	private Date invDcDate;
